Refactor SqlQueryQueueManager
Move QueryQueueRule creation to a separate class
